third_party/WebKit/LayoutTests/typedcssom/cssUnitValue_toMethod.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>

<script>
var EPSILON = 1e-6; // float epsilon

let angleUnits = [
  'deg',
  'rad',
  'grad',
  'turn'
];

let conversionFactors = {
  'deg': {
    'deg': 1,
    'rad': Math.PI / 180,
    'grad': 400 / 360,
    'turn': 1 / 360,
  },
  'rad': {
    'deg': 180 / Math.PI,
    'rad': 1,
    'grad': 200 / Math.PI,
    'turn': 0.5 / Math.PI,
  },
  'grad': {
    'deg': 360 / 400,
    'rad': Math.PI / 200,
    'grad': 1,
    'turn': 1 / 400,
  },
  'turn': {
    'deg': 360,
    'rad': 2 * Math.PI,
    'grad': 400,
    'turn': 1,
  },
}

test(() => {
  let unitValue = new CSSUnitValue(1, 'deg');
  assert_throws(new SyntaxError(), () => {
    unitValue.to('bananas');
  });
}, 'Converting to invalid unit throws');

test(() => {
  let unitValue = new CSSUnitValue(1, 'deg');
  assert_throws(new TypeError(), () => {
    unitValue.to('px');
  });
}, 'Converting to a non-compatible unit throws');

for (let unit of angleUnits) {
  for (let toUnit of angleUnits) {
    test(() => {
      let unitValue = new CSSUnitValue(1, unit);
      let result = unitValue.to(toUnit);
      assert_approx_equals(
          result.value, conversionFactors[unit][toUnit], EPSILON);
      assert_equals(result.unit, toUnit);
    }, 'Converting angle unit ' + unit + ' to ' + toUnit);
  }
}

</script>
